+++ lib/active_record/connection_adapters/sqlserver_adapter.rb
@@ -443,6 +443,8 @@ module ActiveRecord
 
       def translate_exception(e, message:, sql:, binds:)
         case message
+        when /(SQL Server client is not connected)|(failed to execute statement)/i
+          ConnectionNotEstablished.new(message)
         when /(cannot insert duplicate key .* with unique index) | (violation of unique key constraint)/i
           RecordNotUnique.new(message, sql: sql, binds: binds)
         when /(conflicted with the foreign key constraint) | (The DELETE statement conflicted with the REFERENCE constraint)/i